Switch gAskpass to command-line handler
Removed file-open handler from gAskpass and switched it to the generic
command-line handler. Moved argc/argv handling code out of main() and
in to a dedicated callback function.

This change should address issue #4: gaskpass window not displayed,
reported by hrotkogabor.

Using ssh with password authentication makes me sad. That is a separate
issue, and this change does not address it at all.