gstm: Use setsid to launch gstm from launcher
On Ubuntu 18.04 (and possibly other distributions), applications
may incorrectly latch on to the X.org's console as the launching terminal.
In gstm's case, this results in child ssh processes also inheriting a
working terminal, which causes SSH_ASKPASS to be ignored and passwords
are requested on the console rather than via a gui. Launching with
setsid apparently fixes this.