gstm: Replace GtkStatusIcon with AppIndicator
Replaced depreceated GtkStatusIcon with libappindicator.
Replaced depreceated menu functions with Gtk Box-based work-around, allowing
icons to be kept in the docklet menu.

Thanks go to Alex Guzman (https://github.com/reanimus) for finishing my
initial work in this direction and for showing there was user interest in
making the switch.