gstm: Update AppIndicator menu.
Update AppIndicator menu when a tunnel is added, removed, started or
stopped.

The old docklet (GtkStatusIcon) had it's menu re-generated every time
the docklet recieved a click event, ensuring it was always up-to-date.
With AppIndicator, the application gets no notification whatsoever when the
docklet menu is opened, meaning any event or status change within the
application that would need to be reflected on the menu has to
explicitly update the menu at the code that caused the event or status
change to occur.