gstm: Read ~/.ssh/config for preset Host entries.
gstm will now scan ~/.ssh/config and list any non-* Host entries (ie,
"Host *" is ignored, "Host whatever" is not) as available "Presets" to
be selected in a drop-down.

Selecting a Preset will set Host to the Preset and blank User, Port, and
Privkey.

If a preset is selected, it is assumed any and all configuration, beyond
the port forwarding/redirects, will be handled by the user in
~/.ssh/config.