Make single-instance behavior work correctly
The way GApplication works, whenever an app starts up, it will send the
startup and activate signals and call the associated callbacks for your
app. Later, if the app is unique (which is the default), if you open the
app again, it will send the running instance an activate signal.

They way this works right now with gstm, it will result in the main
instance reinitializing, losing track of any running tunnels, then
generating another main window, duplicating any existing ones.

This makes the activation keep track of whether or not it's the initial
activation, simply presenting the main window if it isn't. I had some
crashes when I tried to move this code to the startup function, so I
figure this is the most straightforward way to handle it.